Prepare to release version 1.0.0 (#23)
Changes:
* Add a standard .gitignore file
* Set the Julia compatibility version to ^1.0
* Remove the REQUIRE file
* Replace testing on 0.7 with 1.2
* Set the package version to 1.0.0.

Since Reexport is such a stable package, it makes sense for it to
declare itself to be at version 1.0. That way packages which depend on
it can specify their version requirement as `Reexport = "1"` without
needing to add a new 0.x version of Reexport to their `[compat]` every
time one is released.