Fix use of reexport inside of baremodules
As described in issue 32, the generated code from the macro included
references to `Base`, which notably is not defined in `baremodule`s. We
can get around this by moving the part that fetches and filters module
names into its own function, ensure that gets used as a `GlobalRef` by
the macro, and swap the `eval` for `Core.eval`, which is available in
`baremodule`s.